Implement assign and eliminate strategy.
As described by Peter Norvig, the board tracks the current value for
each square, or a list of possible values (when empty.)

This especially improves the speed of checking if a value is possible
for a square, and propagating the decision of setting it.

Brings the speed of the test suite against all 50 easy problems and 95
hard problems from 128s down to ~11s.